fix: Address review feedback on workflow CLI commands
- Add workflow commands to README CLI table (amsterdam #1)
- Move inline tests to separate workflow_tests.rs file (amsterdam #2)
- Document channel_workflows and new RPC methods in architecture.md (amsterdam #3)
