+++ docs/architecture.md
@@ -734,7 +734,7 @@ Five RPC methods in `rpc_workflow.rs` manage workflow state and channel assignme
 
 - **`workflow.get_state`** — reads from the in-memory `workflow_state` map for a channel, optionally scoped to a plugin key. Returns `null` when absent.
 - **`workflow.set_state`** — updates the in-memory state and persists to `daemon-state.json`. With a `plugin` key, merges at that key; without, replaces the entire channel state. Concurrent writes are serialized by the `persistent_state` Mutex.
-- **`workflow.assign`** — maps a channel to a named workflow (stored in `DaemonPersistentState::channel_workflows`). Multiple channels can share the same workflow.
+- **`workflow.assign`** — maps a channel to a named workflow (stored in `DaemonPersistentState::channel_workflows`). Validates the workflow exists before saving. Multiple channels can share the same workflow.
 - **`workflow.unassign`** — removes a channel's workflow assignment, reverting to daemon defaults.
 - **`workflow.list`** — returns available workflows discovered from `~/.midtown/projects/<project>/workflows/` (directories containing `workflow.py`) and current channel→workflow assignments.
