+++ src/daemon/state_tests.rs
@@ -1148,7 +1148,6 @@ fn apply_gc_combined_operations() {
 
 // ── channel_workflows tests ───────────────────────────────────────────
 
-
 #[test]
 fn test_channel_workflow_assignment_roundtrip() {
     let mut state = DaemonPersistentState::default();
